refactor(chat): reorganize chat components and enhance functionality
- Updated import paths for ChatComposer and ChatPage components to improve structure.
- Introduced new components: ChatAppHeader, ChatAutoScrollButton, and MarkdownContent for better modularity.
- Added support for code blocks and markdown rendering in messages, enhancing message display capabilities.
- Implemented attachment handling and improved user interactions within the chat interface.
- Refactored existing components to streamline functionality and improve code readability.
