+++ features/chat/components/chat-auto-scroll-button.tsx
