+++ features/chat/components/chat-page.tsx
@@ -1,10 +1,10 @@
 'use client';
 
-import ChatComposer from './chat-composer';
+import ChatComposer from '../composer/chat-composer';
 import ChatAutoScrollButton from './chat-auto-scroll-button';
 import ChatAppHeader from './chat-app-header';
 import Messages from './messages';
-import useChatPage from './use-chat-page';
+import useChatPage from '../hooks/use-chat-page';
 
 type Props = {
   chatId: string;
