+++ features/chat/hooks/use-auto-scroll.ts
